requirements.txt
asgiref==3.3.4
Django==3.2
django-ckeditor==6.2.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-sms==0.5.0
djangorestframework==3.12.4
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
